PushNotificationTarget のインターフェース抽出
